+++ frontend/components/configurations/LocationMap.vue
@@ -2,11 +2,12 @@
 Web client of the Sensor Management System software developed within the
 Helmholtz DataHub Initiative by GFZ and UFZ.
 
-Copyright (C) 2020, 2021
+Copyright (C) 2020 - 2023
 - Nils Brinckmann (GFZ, nils.brinckmann@gfz-potsdam.de)
 - Marc Hanisch (GFZ, marc.hanisch@gfz-potsdam.de)
 - Tobias Kuhnert (UFZ, tobias.kuhnert@ufz.de)
 - Erik Pongratz (UFZ, erik.pongratz@ufz.de)
+- Maximilian Schaldach (UFZ, maximilian.schaldach@ufz.de)
 - Helmholtz Centre Potsdam - GFZ German Research Centre for
   Geosciences (GFZ, https://www.gfz-potsdam.de)
 - Helmholtz Centre for Environmental Research GmbH - UFZ
@@ -36,15 +37,30 @@ permissions and limitations under the Licence.
   <div id="map-wrap" style="height: 300px">
     <no-ssr>
       <l-map
+        ref="map"
         :zoom="10"
-        :center="currentPosition"
+        :center="currentMapCenter"
         style="z-index:0"
-        @click="setCoordinate"
+        @click="setCoordinatesByMapClick"
         @ready="onReady"
         @locationfound="onLocationFound"
       >
         <l-tile-layer url="https://{s}.tile.osm.org/{z}/{x}/{y}.png" />
         <l-marker v-if="location" :lat-lng="location" />
+        <v-btn
+          v-if="!readonly"
+          id="set-current-location-button"
+          fab
+          small
+          color="white"
+          :disabled="isLocating"
+          @click.stop="setCoordinatesByCurrentLocation"
+        >
+          <v-icon v-if="!isLocating">
+            mdi-map-marker-account
+          </v-icon>
+          <v-progress-circular v-else indeterminate />
+        </v-btn>
       </l-map>
     </no-ssr>
   </div>
@@ -64,27 +80,47 @@ export default class LocationMap extends Vue {
 
   @Prop({ default: false, type: Boolean }) readonly readonly!: boolean
 
-  private currentPosition: LatLng = new LatLng(52, 12)
-  private currentPositionIntializedByLocation = false
+  private currentPosition: LatLng|null = null
+  private isLocating: Boolean = false
+  private defaultMapCenter: LatLng = new LatLng(52, 12)
+  private currentMapCenter: LatLng = this.location ?? this.defaultMapCenter
+  private currentMapCenterInitializedByLocation = false
 
   created () {
     if (this.location !== null) {
       this.currentPosition = this.location
-      this.currentPositionIntializedByLocation = true
+      this.currentMapCenterInitializedByLocation = true
     }
   }
 
-  setCoordinate (event: LeafletMouseEvent) {
+  setCoordinatesByMapClick (event: LeafletMouseEvent) {
     if (this.readonly) {
       return
     }
 
+    this.setCoordinates(event.latlng)
+  }
+
+  setCoordinatesByCurrentLocation () {
+    if (!this.currentPosition) {
+      if (!this.mapObject) { return }
+      this.mapObject.locate()
+      this.isLocating = true
+      return
+    }
+
+    this.setCoordinates(this.currentPosition)
+  }
+
+  setCoordinates (latlng: LatLng) {
     const newValue = StationaryLocation.createFromObject({
-      latitude: event.latlng.lat,
-      longitude: event.latlng.lng,
-      elevation: event.latlng.alt || null
+      latitude: latlng.lat,
+      longitude: latlng.lng,
+      elevation: latlng.alt || null
     })
     this.$emit('input', newValue)
+
+    this.currentMapCenter = latlng
   }
 
   get location (): LatLng|null {
@@ -94,17 +130,39 @@ export default class LocationMap extends Vue {
     return null
   }
 
+  get mapObject (): Map|null {
+    const mapRef: any|null = this.$refs.map
+    const mapObject: Map|null = mapRef?.mapObject ?? null
+    return mapObject
+  }
+
+  locatePosition () {
+    if (!this.mapObject) { return }
+    this.mapObject.locate()
+  }
+
   onReady (mapObject: Map) {
-    // Only locate when no coordinates were set
-    if (this.location === null) {
-      mapObject.locate()
-    }
+    mapObject.on('locationerror', this.onLocationError)
+    mapObject.on('locationfound', this.onLocationFound)
   }
 
   onLocationFound (location: LocationEvent) {
-    if (this.location === null) {
+    this.isLocating = false
+
+    if (this.currentPosition === null) {
       this.currentPosition = location.latlng
+      this.setCoordinatesByCurrentLocation()
     }
+
+    // Only recenter when no marker is set
+    if (this.location === null) {
+      this.currentMapCenter = location.latlng
+    }
+  }
+
+  onLocationError () {
+    this.isLocating = false
+    this.$store.commit('snackbar/setError', 'Location could not be determined.')
   }
 
   @Watch('location')
@@ -122,13 +180,20 @@ export default class LocationMap extends Vue {
     // However, if we use the component just to look at the location (readonly mode),
     // then an updated location means that we should show another static location.
     // In this case we want to update the current position by the given location.
-    if (location !== null && (!this.currentPositionIntializedByLocation || this.readonly)) {
-      this.currentPosition = location
-      this.currentPositionIntializedByLocation = true
+    if (location !== null && (!this.currentMapCenterInitializedByLocation || this.readonly)) {
+      this.currentMapCenter = location
+      this.currentMapCenterInitializedByLocation = true
     }
   }
 }
 </script>
 
 <style scoped>
+#set-current-location-button {
+  z-index: 500;
+  position: absolute;
+  right: 0;
+  bottom: 1.5em;
+  margin: 1em;
+}
 </style>